[cache] refactor setup and teardown procedures ensuring tables always deleted

+++ tests/utest/cache/unit/CacheTest.php
@@ -64,8 +64,10 @@ class CacheTest extends CacheTestCase
 
     protected function tearDownFixtures()
     {
-        $this->connection->getSchemaManager()
-            ->dropTable($this->tableName);
+        if($this->connection->getSchemaManager()->tablesExist($this->tableName)){
+            $this->connection->getSchemaManager()
+                ->dropTable($this->tableName);
+        }
     }
 
     public function testCacheSimple()
